+++ R/utils.R
@@ -1,7 +1,7 @@
 #' Get the terminal node id of a RandomForest Object
 #' 
 #' Extracts for each observation and for each tree in the forest the terminal 
-#' node id.
+#' node id. The index of terminal nodes are starting with 1, e.g., the root node has id 1
 #'
 #' @param x a data.frame
 #' @param rfObject \code{ranger} object
@@ -12,10 +12,10 @@
 #' @examples
 #' library(ranger)
 #' rf.fit <- ranger(Species ~ ., data = iris, num.trees = 5, write.forest = TRUE)
-#' dfNodes <- terminalNodeIDs(iris[, -5], rf.fit)
+#' dfNodes <- terminalNodes(iris[, -5], rf.fit)
 #' 
 #' @export
-terminalNodeIDs <- function(x, rfObject) {
+terminalNodes <- function(x, rfObject) {
   testthat::expect_is(rfObject, "ranger")
   testthat::expect_false(object = is.null(rfObject$forest), 
                          info   = "Ranger object does not contain a forest.")
@@ -28,7 +28,8 @@ terminalNodeIDs <- function(x, rfObject) {
                        splitVarIds = rfObject$forest$split.varIDs[[tree]])
   }, simplify = F)
   res <- do.call(cbind, res)
-  as.matrix(res)
+  # switch from 0-index to 1-index
+  as.matrix(res) + 1
 }
 
 
@@ -50,16 +51,18 @@ terminalNodeIDs <- function(x, rfObject) {
 #' forestMat <- forestToMatrix(rf.fit)
 #' 
 #' @export
-forestToMatrix <- function(rfObject) {
+ranger_forests_to_matrix <- function(rfObject) {
   res <- sapply(1:rfObject$num.trees, function(t) {
     len <- length(rfObject$forest$child.nodeIDs[[t]][[1]])
     data.frame(t   = rep(t, len), 
                n   = seq_len(len),
                id1 = rfObject$forest$child.nodeIDs[[t]][[1]],
-               id2 = rfObject$forest$child.nodeIDs[[t]][[2]])
+               id2 = rfObject$forest$child.nodeIDs[[t]][[2]],
+               split_id = rfObject$forest$split.varIDs[[t]])
   }, simplify = F)
   res <- do.call(rbind, res)
-  as.matrix(res)
+  res %>% 
+    as.matrix()
 }
 
 
@@ -78,4 +81,14 @@ asDistObject <- function(x, n, method) {
             Upper  = F,
             method = method,
             class  = "dist")
+}
+
+
+#' call a function by character strings
+call_function = function(func_list) {
+  func_name <- func_list$func
+  func_namespace <- func_list$namespace
+  func_args <- func_list$args
+  func_to_call <- get(func_name, envir = as.environment(func_namespace))
+  pryr::do_call(func_to_call, func_args)
 }