+++ NEWS.md
@@ -1,3 +1,19 @@
+## 0.3
+
+### Features
+
+- Add more unit tests
+- Add R6 documentation
+- Generalize the RandomForest model to classification and regression
+- Update vignettes
+
+### Bugfixes
+
+- Fix ranger package to version 0.11.2
+- Terminal nodes were wrongly calculated
+- Fix Armadillo deprecation warnings
+
+
 ## 0.22
 
 ### Bugfixes